# infra/gateway/openapi-google.yaml
swagger: "2.0"
info:
  title: innerpeace-api
  version: "1.0.0"
schemes: ["https"]
produces: ["application/json"]
securityDefinitions:
  app_jwt:
    type: oauth2
    flow: implicit
    authorizationUrl: ""
    x-google-issuer: "https://innerpeace.app"
    x-google-jwks_uri: "https://innerpeace.app/.well-known/jwks.json"
    x-google-audiences: "innerpeace-app"
    x-google-jwt-locations:
      - header: "x-app-jwt"
      - header: "Authorization"
        value_prefix: "Bearer "

  google_id_token:
    type: oauth2
    flow: implicit
    authorizationUrl: ""
    x-google-issuer: "https://accounts.google.com"
    x-google-jwks_uri: "https://www.googleapis.com/oauth2/v3/certs"
    x-google-audiences: "379484922687-q7ge8kg89o0vi1gn1lkjaqciu8e8e3eb.apps.googleusercontent.com"
paths:
...
  /api/book:
    post:
      operationId: createBooking
      security:
        - app_jwt: []
        - google_id_token: []
      x-google-backend:
        address: "https://innerpeace-api-379484922687.europe-west2.run.app"
        path_translation: APPEND_PATH_TO_ADDRESS
        jwt_audience: "https://innerpeace-api-379484922687.europe-west2.run.app"
      responses:
        "200": { description: "OK" }
...